fixed speed on domyosbike #203

+++ src/domyosbike.cpp
@@ -334,7 +334,7 @@ void domyosbike::characteristicChanged(const QLowEnergyCharacteristic &character
 
 double domyosbike::GetSpeedFromPacket(QByteArray packet)
 {
-    uint16_t convertedData = (packet.at(6) << 8) | packet.at(7);
+    uint16_t convertedData = (packet.at(6) << 8) | ((uint8_t)packet.at(7));
     double data = (double)convertedData / 10.0f;
     return data;
 }
